// user-service/src/grpc/generated/auth-user.ts
export interface userRegistrationRequest {
    name: string;
    email: string;
    hashedPassword: string;
}

export interface userRegistrationResponse {
    userID: string;
}

export interface getUserIdPassRequest {
    email: string;
}

export interface getUserIdPassResponse {
    userID: string;
    hashedPassword: string;
}

function createBaseuserRegistrationRequest(): userRegistrationRequest {
    return { name: "", email: "", hashedPassword: "" };
}

export const userRegistrationRequest: MessageFns<userRegistrationRequest> = {
    encode(message: userRegistrationRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
        if (message.name !== "") {
            writer.uint32(10).string(message.name);
        }
        if (message.email !== "") {
            writer.uint32(18).string(message.email);
        }
        if (message.hashedPassword !== "") {
            writer.uint32(26).string(message.hashedPassword);
        }
        return writer;
    },

    decode(input: BinaryReader | Uint8Array, length?: number): userRegistrationRequest {
        const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
        const end = length === undefined ? reader.len : reader.pos + length;
        const message = createBaseuserRegistrationRequest();
        while (reader.pos < end) {
            const tag = reader.uint32();
            switch (tag >>> 3) {
                case 1: {
                    if (tag !== 10) {
                        break;
                    }

                    message.name = reader.string();
                    continue;
                }
                case 2: {
                    if (tag !== 18) {
                        break;
                    }

                    message.email = reader.string();
                    continue;
                }
                case 3: {
                    if (tag !== 26) {
                        break;
                    }

                    message.hashedPassword = reader.string();
                    continue;
                }
            }
            if ((tag & 7) === 4 || tag === 0) {
                break;
            }
            reader.skip(tag & 7);
        }
        return message;
    },

    fromJSON(object: any): userRegistrationRequest {
        return {
            name: isSet(object.name) ? globalThis.String(object.name) : "",
            email: isSet(object.email) ? globalThis.String(object.email) : "",
            hashedPassword: isSet(object.hashedPassword) ? globalThis.String(object.hashedPassword) : "",
        };
    },

    toJSON(message: userRegistrationRequest): unknown {
        const obj: any = {};
        if (message.name !== "") {
            obj.name = message.name;
        }
        if (message.email !== "") {
            obj.email = message.email;
        }
        if (message.hashedPassword !== "") {
            obj.hashedPassword = message.hashedPassword;
        }
        return obj;
    },

    create<I extends Exact<DeepPartial<userRegistrationRequest>, I>>(base?: I): userRegistrationRequest {
        return userRegistrationRequest.fromPartial(base ?? ({} as any));
    },
    fromPartial<I extends Exact<DeepPartial<userRegistrationRequest>, I>>(object: I): userRegistrationRequest {
        const message = createBaseuserRegistrationRequest();
        message.name = object.name ?? "";
        message.email = object.email ?? "";
        message.hashedPassword = object.hashedPassword ?? "";
        return message;
    },
};

function createBaseuserRegistrationResponse(): userRegistrationResponse {
    return { userID: "" };
}

export const userRegistrationResponse: MessageFns<userRegistrationResponse> = {
    encode(message: userRegistrationResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
        if (message.userID !== "") {
            writer.uint32(10).string(message.userID);
        }
        return writer;
    },

    decode(input: BinaryReader | Uint8Array, length?: number): userRegistrationResponse {
        const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
        const end = length === undefined ? reader.len : reader.pos + length;
        const message = createBaseuserRegistrationResponse();
        while (reader.pos < end) {
            const tag = reader.uint32();
            switch (tag >>> 3) {
                case 1: {
                    if (tag !== 10) {
                        break;
                    }

                    message.userID = reader.string();
                    continue;
                }
            }
            if ((tag & 7) === 4 || tag === 0) {
                break;
            }
            reader.skip(tag & 7);
        }
        return message;
    },

    fromJSON(object: any): userRegistrationResponse {
        return { userID: isSet(object.userID) ? globalThis.String(object.userID) : "" };
    },

    toJSON(message: userRegistrationResponse): unknown {
        const obj: any = {};
        if (message.userID !== "") {
            obj.userID = message.userID;
        }
        return obj;
    },

    create<I extends Exact<DeepPartial<userRegistrationResponse>, I>>(base?: I): userRegistrationResponse {
        return userRegistrationResponse.fromPartial(base ?? ({} as any));
    },
    fromPartial<I extends Exact<DeepPartial<userRegistrationResponse>, I>>(object: I): userRegistrationResponse {
        const message = createBaseuserRegistrationResponse();
        message.userID = object.userID ?? "";
        return message;
    },
};

function createBasegetUserIdPassRequest(): getUserIdPassRequest {
    return { email: "" };
}

export const getUserIdPassRequest: MessageFns<getUserIdPassRequest> = {
    encode(message: getUserIdPassRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
        if (message.email !== "") {
            writer.uint32(10).string(message.email);
        }
        return writer;
    },

    decode(input: BinaryReader | Uint8Array, length?: number): getUserIdPassRequest {
        const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
        const end = length === undefined ? reader.len : reader.pos + length;
        const message = createBasegetUserIdPassRequest();
        while (reader.pos < end) {
            const tag = reader.uint32();
            switch (tag >>> 3) {
                case 1: {
                    if (tag !== 10) {
                        break;
                    }

                    message.email = reader.string();
                    continue;
                }
            }
            if ((tag & 7) === 4 || tag === 0) {
                break;
            }
            reader.skip(tag & 7);
        }
        return message;
    },

    fromJSON(object: any): getUserIdPassRequest {
        return { email: isSet(object.email) ? globalThis.String(object.email) : "" };
    },

    toJSON(message: getUserIdPassRequest): unknown {
        const obj: any = {};
        if (message.email !== "") {
            obj.email = message.email;
        }
        return obj;
    },

    create<I extends Exact<DeepPartial<getUserIdPassRequest>, I>>(base?: I): getUserIdPassRequest {
        return getUserIdPassRequest.fromPartial(base ?? ({} as any));
    },
    fromPartial<I extends Exact<DeepPartial<getUserIdPassRequest>, I>>(object: I): getUserIdPassRequest {
        const message = createBasegetUserIdPassRequest();
        message.email = object.email ?? "";
        return message;
    },
};

function createBasegetUserIdPassResponse(): getUserIdPassResponse {
    return { userID: "", hashedPassword: "" };
}

export const getUserIdPassResponse: MessageFns<getUserIdPassResponse> = {
    encode(message: getUserIdPassResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
        if (message.userID !== "") {
            writer.uint32(10).string(message.userID);
        }
        if (message.hashedPassword !== "") {
            writer.uint32(18).string(message.hashedPassword);
        }
        return writer;
    },

    decode(input: BinaryReader | Uint8Array, length?: number): getUserIdPassResponse {
        const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
        const end = length === undefined ? reader.len : reader.pos + length;
        const message = createBasegetUserIdPassResponse();
        while (reader.pos < end) {
            const tag = reader.uint32();
            switch (tag >>> 3) {
                case 1: {
                    if (tag !== 10) {
                        break;
                    }

                    message.userID = reader.string();
                    continue;
                }
                case 2: {
                    if (tag !== 18) {
                        break;
                    }

                    message.hashedPassword = reader.string();
                    continue;
                }
            }
            if ((tag & 7) === 4 || tag === 0) {
                break;
            }
            reader.skip(tag & 7);
        }
        return message;
    },

    fromJSON(object: any): getUserIdPassResponse {
        return {
            userID: isSet(object.userID) ? globalThis.String(object.userID) : "",
            hashedPassword: isSet(object.hashedPassword) ? globalThis.String(object.hashedPassword) : "",
        };
    },

    toJSON(message: getUserIdPassResponse): unknown {
        const obj: any = {};
        if (message.userID !== "") {
            obj.userID = message.userID;
        }
        if (message.hashedPassword !== "") {
            obj.hashedPassword = message.hashedPassword;
        }
        return obj;
    },

    create<I extends Exact<DeepPartial<getUserIdPassResponse>, I>>(base?: I): getUserIdPassResponse {
        return getUserIdPassResponse.fromPartial(base ?? ({} as any));
    },
    fromPartial<I extends Exact<DeepPartial<getUserIdPassResponse>, I>>(object: I): getUserIdPassResponse {
        const message = createBasegetUserIdPassResponse();
        message.userID = object.userID ?? "";
        message.hashedPassword = object.hashedPassword ?? "";
        return message;
    },
};

export type AuthUserCommunicationService = typeof AuthUserCommunicationService;
export const AuthUserCommunicationService = {
    userRegistration: {
        path: "/authUserCommunication.AuthUserCommunication/userRegistration",
        requestStream: false,
        responseStream: false,
        requestSerialize: (value: userRegistrationRequest): Buffer => Buffer.from(userRegistrationRequest.encode(value).finish()),
        requestDeserialize: (value: Buffer): userRegistrationRequest => userRegistrationRequest.decode(value),
        responseSerialize: (value: userRegistrationResponse): Buffer => Buffer.from(userRegistrationResponse.encode(value).finish()),
        responseDeserialize: (value: Buffer): userRegistrationResponse => userRegistrationResponse.decode(value),
    },
    getUserIdPass: {
        path: "/authUserCommunication.AuthUserCommunication/getUserIdPass",
        requestStream: false,
        responseStream: false,
        requestSerialize: (value: getUserIdPassRequest): Buffer => Buffer.from(getUserIdPassRequest.encode(value).finish()),
        requestDeserialize: (value: Buffer): getUserIdPassRequest => getUserIdPassRequest.decode(value),
        responseSerialize: (value: getUserIdPassResponse): Buffer => Buffer.from(getUserIdPassResponse.encode(value).finish()),
        responseDeserialize: (value: Buffer): getUserIdPassResponse => getUserIdPassResponse.decode(value),
    },
} as const;
// ...
export type DeepPartial<T> = T extends Builtin
    ? T
    : T extends globalThis.Array<infer U>
    ? globalThis.Array<DeepPartial<U>>
    : T extends ReadonlyArray<infer U>
    ? ReadonlyArray<DeepPartial<U>>
    : T extends {}
    ? { [K in keyof T]?: DeepPartial<T[K]> }
    : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
    return value !== null && value !== undefined;
}

export interface MessageFns<T> {
    encode(message: T, writer?: BinaryWriter): BinaryWriter;
    decode(input: BinaryReader | Uint8Array, length?: number): T;
    fromJSON(object: any): T;
    toJSON(message: T): unknown;
    create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
    fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}